changing setup scripts

+++ docs/setup_linux.sh
@@ -32,29 +32,41 @@ python3 -m pip install -U dataclasses==0.6
 python3 -m pip install -U -r requirements.txt
 pre-commit && pre-commit install
 
-if lspci -vnnn | perl -lne 'print if /^\d+\:.+(\[\S+\:\S+\])/' | grep -q NVIDIA;
-then
-    echo "GPU found."
-    python3 -m pip install -U -r requirements-gpu.txt
-    pushd $(mktemp -d)
-    rm -rf apex
-    git clone --recursive https://github.com/NVIDIA/apex
-    pushd apex
-    export TORCH_CUDA_ARCH_LIST="6.0;6.1;6.2;7.0;7.5"
-    git checkout 1ff54b8fed441c39dac181091b44fecdca31a403
-    CUDA_HOME="/usr/local/cuda-10.2" pip install -v --no-cache-dir \
-        --global-option="--cpp_ext" --global-option="--cuda_ext" ./
-    popd
-    popd
-else
+IS_MAC=$(uname -a)
+
+if [[ $IS_MAC =~ "Darwin" ]]; then
     echo "No GPU found."
-    python3 -m pip install -U -r requirements-cpu.txt
+    python3 -m pip install -U torch torchvision
+else
+    if lspci -vnnn | perl -lne 'print if /^\d+\:.+(\[\S+\:\S+\])/' | grep -q NVIDIA;
+    then
+        echo "GPU found."
+        python3 -m pip install -U -r requirements-gpu.txt
+        pushd $(mktemp -d)
+        rm -rf apex
+        git clone --recursive https://github.com/NVIDIA/apex
+        pushd apex
+        export TORCH_CUDA_ARCH_LIST="6.0;6.1;6.2;7.0;7.5"
+        git checkout 1ff54b8fed441c39dac181091b44fecdca31a403
+        CUDA_HOME="/usr/local/cuda-10.2" pip install -v --no-cache-dir \
+            --global-option="--cpp_ext" --global-option="--cuda_ext" ./
+        popd
+        popd
+    else
+        python3 -m pip install -U -r requirements-cpu.txt
+    fi
 fi
 
 # Get bazelisk
 pushd $(mktemp -d)
-wget -O bazel https://github.com/bazelbuild/bazelisk/releases/download/v1.7.1/bazelisk-linux-amd64
-chmod 755 bazel
+if [[ $IS_MAC =~ "Darwin" ]]; then
+    echo "Downloading Bazelisk for Darwin"
+    curl -fL -o bazel https://github.com/bazelbuild/bazelisk/releases/download/v1.7.2/bazelisk-darwin-amd64
+else
+    curl -fL -o bazel https://github.com/bazelbuild/bazelisk/releases/download/v1.7.1/bazelisk-linux-amd64
+fi
+
+chmod +x bazel
 sudo mv bazel /usr/local/bin
 popd
